+++ PokeTactics.Api/Endpoints/PokemonEndpoints.cs
@@ -1,5 +1,6 @@
+using PokeTactics.Contracts.Common.Requests;
+using PokeTactics.Contracts.Common.Responses;
 using PokeTactics.Contracts.Pokemon.Responses;
-using PokeTactics.Core.Definitions.Dtos;
 using PokeTactics.Services.Facade;
 
 namespace PokeTactics.Api.Endpoints;
