+++ PokeTactics.Contracts/Common/Responses/KeysetPaginationResponse.cs
@@ -1,4 +1,4 @@
-namespace PokeTactics.Core.Definitions.Dtos;
+namespace PokeTactics.Contracts.Common.Responses;
 
 public record KeysetPaginationResponse<T>(
     IEnumerable<T> Items,
