+++ PokeTactics.Services/Facade/Impl/PokemonService.cs
@@ -1,5 +1,6 @@
+using PokeTactics.Contracts.Common.Requests;
+using PokeTactics.Contracts.Common.Responses;
 using PokeTactics.Contracts.Pokemon.Responses;
-using PokeTactics.Core.Definitions.Dtos;
 using PokeTactics.Core.Entities;
 using PokeTactics.Core.Interfaces;
 using PokeTactics.Services.Mappers;
@@ -17,12 +18,12 @@ public class PokemonService : IPokemonService
 
     public async Task<KeysetPaginationResponse<PokemonDto>> Find(KeysetPaginationRequest request)
     {
-        return await ExecuteFind(() => _unitOfWork.PokemonDao.Find(request));
+        return await ExecuteFind(() => _unitOfWork.PokemonDao.Find(request.PageSize, request.LastId));
     }
 
     public async Task<KeysetPaginationResponse<PokemonDto>> FindDeep(KeysetPaginationRequest request)
     {
-        return await ExecuteFind(() => _unitOfWork.PokemonDao.FindDeep(request));
+        return await ExecuteFind(() => _unitOfWork.PokemonDao.FindDeep(request.PageSize, request.LastId));
     }
 
     private static async Task<KeysetPaginationResponse<PokemonDto>> ExecuteFind(Func<Task<ICollection<Pokemon>>> find)
